Settings files where ignored
When compiling this project from oustide the LF the settings files
should be used to change where the dependencies and plugins should
be fetched. The `MVN` environment variable only contains the path
to the mvn executable and is lacking of the parameters to take the
global settings file and settings file for the project into account.

Change-Id: Ice43e8e297cd12fce317941864a8add759e99efd
Signed-off-by: Anaël Closson <ac2550@intl.att.com>