Reduce image size of containers
CIA project goal is to reduce image size of container and make the
container multi architecture.

Portal images are already multi architecure ready and portal-wms and
portal-db are also done in an efficient way.

Portal-app and portal-sdk although can be shrinked, using multi stage
docker builds.

This commit allow a noticeable shrink for both of them:

from 708Mb to 298Mb for portal-app and from 388Mb to 257Mb for
portal-sdk.

Change-Id: Id5cca8b78d0d9ee1724259dab40b7ccff910c628
Issue-ID: INT-939
Issue-ID: INT-922
Issue-ID: INT-881
Issue-ID: INT-925
Issue-ID: INT-936
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>